update content_lenght to c_long kind

+++ src/http/http_response.f90
@@ -11,7 +11,7 @@ module http_response
     type :: response_type
         character(len=:), public, allocatable :: url, content, method, err_msg, header_string
         integer, public :: status_code = 0
-        integer(kind=8), public :: content_length = 0
+        integer(kind=c_long), public :: content_length = 0
         logical, public :: ok = .true.
         type(fhash_tbl_t) :: header
     end type response_type
